+++ packages/cinterop/src/nativeDarwin/kotlin/io/realm/kotlin/internal/interop/RealmInterop.kt
@@ -2002,7 +2002,7 @@ actual object RealmInterop {
                 realm_wrapper.realm_app_get_all_users(
                     app.cptr(),
                     null,
-                    0,
+                    0UL,
                     capacityCount.ptr
                 )
             )